backend/requirements.txt
flask
flask-cors
opencv-python
numpy
dlib
face-recognition
pymongo
mysql-connector-python
werkzeug